WIP TTLeagues TeamSummary page

+++ src/Smab.TTInfo.TTLeagues/Models/Match.cs
@@ -2,12 +2,12 @@
 
 internal sealed class Match
 {
-	public long Id { get; set; }
+	public int Id { get; set; }
 	public required MatchTeamInfo Home { get; set; }
 	public required MatchTeamInfo Away { get; set; }
 	public DateTime? Date { get; set; }
 	public DateTime? Time { get; set; }
-	public long? Week { get; set; }
+	public int? Week { get; set; }
 	public required string Name { get; set; }
 	public required string Venue { get; set; }
 	public long? GroupingId { get; set; }
